requirements.txt
astroid==2.8.2
click==8.0.1
ecdsa==0.17.0
Flask==1.0.2
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.4.0
future==0.18.2
gunicorn==20.1.0
importlib-metadata==4.8.1
isort==4.3.21
itsdangerous==2.0.1
Jinja2==3.0.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
psycopg2==2.9.1
pycryptodome==3.3.1
pylint==2.3.1
python-jose-cryptodome==1.3.2
six==1.16.0
SQLAlchemy==1.3.3
typed-ast==1.4.3
typing-extensions==3.10.0.2
Werkzeug==2.0.2
wrapt==1.12.1
zipp==3.5.0
